P518-022 Fix problem with Why3 not recognizing previous Coq proof
After Why3 has created a new Coq proof file, it tries to match it to the
existing registered proof file for that VC. This did not work on the
generated proof files due to the preprocessing option used, which stripped
blank lines, which are needed by Why3 printer/parser for Coq.

* include/Makefile
Change generation of proof files to avoid stripping blank lines.

* include/proof/sessions/spark-integer_arithmetic_lemmas/why3session.xml
* ../proof/sessions/spark-long_integer_arithmetic_lemmas/why3session.xml
* include/proof/sessions/spark-mod32_arithmetic_lemmas/why3session.xml
* include/proof/sessions/spark-mod64_arithmetic_lemmas/why3session.xml
New registered proofs.