minor removal of useless code in Parse_Initializes
It does not make sense to call Parse_Initializes without a package entity.
This is now enforced by contract. As a consequence, it did not make sense in
the body to check if the Package is present.

Also: Useless call to Unique_Entity, which was just an identity function, is
now removed. Repeated calls to Abstract_States rewritten. Repeated calls
related to check for Null abstract state rewritten. WHILE loop rewritten
with condition at the end, because the Elist -- if present -- is guaranteed
to be non-empty and the check at the start was useless. Repeated map
searching with Contains followed by Insert rewritten with a single search.
Repeated calls to Direct_Mapping_Id rewritten.

* gnat2why/flow/flow_dependency_maps.ads
(Parse_Initializes): stronger contract

* gnat2why/flow/flow_dependency_maps.adb
(Parse_Initializes): cleanup of useless and repeated code