P523-010 Use absolute path for tests of SPARK lemma library
The variable SPARK_LEMMAS_OBJECT_DIR should be set to an absolute path,
otherwise artefacts are created in the SPARK install. Fix the testsuite
and document that info in the UG.

* docs/ug/gnatprove_by_example/manual_proof.rst
Add info on setting of SPARK_LEMMAS_OBJECT_DIR.

* testsuite/gnatprove/tests/J506-016__scaled_value/test.py
* testsuite/gnatprove/tests/P405-006__arithmetic_from_dafny/test.py
* testsuite/gnatprove/tests/P405-006__arithmetic_lemmas/test.py
* testsuite/gnatprove/tests/P405-006__mod_arithmetic_lemmas/test.py
Update tests to set SPARK_LEMMAS_OBJECT_DIR to absolute path.