minor rewrites in Add_All_Edges
Code for looping over names of variables/subprograms refactored into a
common procedures that take the enumeration literal as an argument.

The "trick" of this rewrite is to create a discriminated record object based
on non-static discriminant: first create a subtype, then an object, and
finally set the object components. This removes the duplication of code for
Proof_Ins, Inputs and Outputs.

Also, if-then-elseif expression rewritten as case expression; useless "use
Global_Graphs" removed.

* gnat2why/flow/flow_generated_globals-phase_2.adb
(Add_All_Edges): rewritten to remove code duplication