minor rewrite of GG ALI parsing
Previously there was a sequence of loops with context-dependent parsing; now
there is a single loop and a single parsing procedure that is only called
for lines with a GG header. This is avoids code repetition and already
allowed to add few sanity checks.

* gnat2why/flow/flow_generated_globals-phase_2.adb
(Load_GG_Info_From_ALI): rewritten

(Corrupted_ALI_File): renamed from Issue_Corrupted_File_Error (the name was
 already repeating information, because "corrupted" already indicates an
 "error" and "issue" refers to side-effect that should be not exposed to the
 caller).