minor separate caches for packages and subprograms
Whenever we asked if a GG entry exists for a given entity name it was clear
from the context if we ask for a package or subprogram/entry/task-type. Now
we keep these disjoint sets of names in separate containers, which is both
clearer (because we reassert what we mean) and faster (because there are
fewer hash collisions and comparisons).

* gnat2why/flow/flow_generated_globals-phase_2.adb
(GG_Exists): renamed from GG_Exists_Cache now contains disjoint sets with
 packages and subprograms.