minor no flow dependency for Is_Mutable_In_Why
The result of Is_Mutable_In_Why depends on no flow analysis; now its code
does not depend on the flow analysis code. This is both simpler and both
efficient. In general, we should not pull in the dependency on flow analysis
unless it is necessary, e.g. where the result can be computed from the GNAT
AST alone.

* gnat2why/why/gnat2why-util.adb
(Is_Mutable_In_Why): dependency on flow analysis removed