P602-029 Get_Flow_Scope rewritten
Previously Get_Flow_Scope was climbing the GNAT AST to find a package and
then was traversing the private declarations to see it started from there.
This was unreliable if the climb starts from a node that does not come from
the source code, but from the front end expansion. For example, here front
end creates an anonymous subtype for the Short_String array type:

   package P is
   private
      type R is record
         Short_String : String (1 .. 10);
      end record;
   end P;

and if we query for its lower bound (1 from the "1 .. 10" expression), we
will not find it when traversing the private declarations.

Now we climb the AST and once we get to Package_Specification node we check
if the previous context was contained in the private declarations list. Such
a check has constant-time complexity, while traversal has linear complexity.

Also, previously the code was very low-level; now it calls
Unique_Defining_Entity where possible.

Also, previously it always called Present on first iteration of the loop,
which was useless because the contract guaranteed that the result is always
true; now the condition for the loop exit is at its end, so no useless check
are done.

* gnat2why/flow/flow_refinement.adb
(Get_Flow_Scope): rewritten

(List_Contains): removed, no longer needed