P523-004 global in variables are initialized in loop
When assuming the dynamic invariant of variables
modified in a loop, assume that parameters and
variable global to the program that have mode in
are initialized.

* gnat2why/spark/spark_util.ads
(Enclosing_Unit): Return the first entry, task,
package, protected object, or subprogram in the scopes
of an entity.

* gnat2why/why/gnat2why-util.ads
(Is_Initialized): Returns True if a global variable is
known to always be initialized in the scope of a unit.

* gnat2why/why/gnat2why-subprograms.adb
(Compute_Dynamic_Property_For_Inputs): Use Is_Initialized
to determine if an input is initialized.

* gnat2why/why/gnat2why-expr-loops.adb
(Transform_Loop_Statement): Use Is_Initialized to
determine if a variable modified by the loop is
initialized.