P607-037 fix crash with po fields used as (in-)out actuals
Previously only objects for which we had a name in gnat2why could be
used as (in-) out actuals. But components of a protected object need to
be treated differently (essentially as record fields).

We simply replace the code that used the name in gnat2why by a call to
Transform_Identifier which already deals with this case. However we need
to disable generation of dereferences, because here we want to have the
mutable variable and not the dereference

* gnat2why-expr.adb
(Computed_Calls): Call Transform_Identifier instead of using name in
  name table