P518-021 frame condition for unmodified fields
For every field of a record part of a variable that
is statically not modified in a loop, we generate
an assumption at the begining of the loop that the
field is preserved.

* gnat2why/why/gnat2why-expr-loops-inv.adb
(Generate_Frame_Condition): Generate a predicate stating
the dynamic property and the preserved fields of variables
modfied in a loop.

* gnat2why/why/gnat2why-expr-loops.adb
(Transform_Loop_Statement): Assume the frame condition
after the loop-invariant.

* gnat2why/why/gnat2why-expr.adb
(Bind_From_Mapping_In_Expr): Loop_Entry_Map may now
contain N_Defining_Identifiers. Properly hanlde them.