minor restore Result_Name after Why3 generation
Global variable Result_Name holds the Why3 name for the 'Result attribute of
a function name. Once the Why3 code for a function has been generated its
meaning is undefined and should not be even accessed. Now this is partly
enforced by restoring its value to Why_Empty in hope that an access will
cause a crash later.

Access function Name_For_Result was supposed to provide an early crash, but
it was not always used and even when it was, it did not detect much because
Result_Name was never restored.

Note: this kind of manipulation of global variable should be really done
with controlled types (just like tampering checks for containers).

* gnat2why/why/gnat2why-expr.adb
(Name_For_Result): removed since not used much anyway