P330-009 flow - Improved info messages and added a test
Fixed some inconsistencies on info messages about initialization. We now
emit a message about initialization for every object which is constituent
of an abstract state (for instance, this was not happening for protected
objects). We suppress the same kind of message for tasks which are not
mentioned on an Initializes aspect.

* flow-analysis.adb
(Find_Use_Of_Uninitialized_Variables): we check if an initialized object,
like a protected object, is a constituent of an abstract state mentioned
in a generated initializes and in this case we output some info.

* flow-control_flow_graph.adb
(Do_Object_Declaration): added a missing check to make sure the entity
considered has been mentioned in a Inizializes aspect.

* testsuite/gnatprove/tests/P330-009__useless_msg_on_task_ob
Added test

* testsuite/gnatprove/tests/
Updated test outputs