P318-019 write GG in the order of analysis
The order of analysis in phase 1 was already fixed, but the result were
still written at random. Now they are written in the order of analysis,
which is what the phase 2 expect.

Also, now we do not collect the partial results in phase 1 and do not need
to re-traverse them when writing down.