rewrite of caching for Entity_Name
Mappings between Entity_Id, Entity_Name and String were implemented with
repeated searching in containers and tentative memory allocation (i.e.
allocating just to compute the hash and check that no allocation is really
necessary).

Now we search and allocate only when necessary. Also, allocation is hidden
in the body of GNATCOLL.Symbols, so we do not manipulate any pointers.
Finally, the new code is (slightly) shorter.

(no-tn-check)