P613-015 Differentiate fields with the same name
Currently, if a hidden component of a tagged type has
the same name as a visible component, the hidden
component is translated as the visible one. When
searching for a component in a record type, we should
return empty if the component is hidden in the record
type.

* gnat2why/spark/spark_util.adb
(Search_Component_By_Name): Do not return a component
with the same name in a tagged type if original
components do not match.