P622-008 flow - implemented check_all mode
Now there is a new check_all mode that reports all messages related to
marking. In this mode flow analysis in run, but error messages that comes
out of it are filtered on severity.

* gnat2why_args.ads
* gnat2why_args.adb
New mode added

* flow_error_messagees.adb
(Error_Msg_Flow): if we are in check_all mode we suppress all messages not
related to marking.

* gnatprove/configuration.ads
* gnatprove/configuration.adb
* gnatprove/gnatprove.adb
New mode added

* tests/P622-008__implementing_check_all
Test added