OA09-017 ignore --report=prover output in inverse_prover
The inverse_prover option will course create a lot of diffs in the
testsuite in connection with the --report=prover option.

We do some postprocessing on the output and the expected test output to
filter the extra output in inverse_prover mode.

also enabling previously disabled tests in inverse-prover testsuite