P627-026 cleanup in Flatten_Variable
Aliasing in flow graphs, which is the main subject of this TN, comes from
wrong handling of concurrent types by Flatten_Variable. This commit merely
clean-ups this routine.

The only semantic change is that if we got something that is not a type, or
whose type entity is one of E_Exception_Type, E_Subprogram_Type or
Incomplete_Kind, then we throw an exception. This gives run-time extra
checks and seems better than returning an unflattened Flow_Id.

Also, the behaviour for abstract states is now explicit; previously it was
silently supported as "something of a misc type."

Also, the contract is now clear about what this routine expects.

* gnat2why/flow/flow_utility.adb
(Flatten_Variable): rewritten