O923-036 no repeated search, asserting and copying
There is no need to assert that a container does not contain an item before
inserting it. (Inserting a duplicate is a run-time error). Also, it caused
repeated search in the container, at least in developer mode (first for
Contains then for Insert). Finally, object of the Info_Maps.Map type might
be expensive to copy. Now it is populated in-place (alternatively it could
be implaced with Move).

* gnat2why/spark/spark_util-types.adb
(Init_Component_Info): rewritten to avoid repeated searching, repeated
 asserting, and inefficient coping.