minor move Why node containers out of tables
Ordered and hashed sets of Why_Node_Id are not related to Why node
tables. Now they are declared in packages where they are used.

Perhaps they were put to Why.Atree.Tables to locate them in some
"common" place; if so, then we should create such a common packages.