P703-003 simplify operations on Why node tables
Operations on Why node tables were implemented with inefficient copying;
now they are implemented as in-place modifications of the AST table.
This (slightly) improves performance in production mode, but reduces it
in development mode because now we need tampering checks.

However, there is no cursor tampering in this package, so it is safe to
suppress tampering checks for Node_Tables and Node_List_Tables packages.
On random test this gives speedup of 1/3.

* gnat2why/why/why-atree-tables.ads
* gnat2why/why/why-atree-tables.adb
Re-implemented to not create local copies of Why AST nodes.