P519-021 proof - add level 5
Add one more proof level (5) which uses a higher step limit.

Add another prover (cvc4_alt) which is used in levels 4 and 5. We do not
run this in the normal testsuite (as it does not add much), only the
tests running under levels. The new prover cvc4_alt is just like cvc4,
except we use a different set of flags that normally make things worse
but sometimes *really* help.

Add a customer test where this new mode manages to prove everything.

Updated user manual with the new level and prover (cvc4_alt).