P616-010 correct wrong interpretation of exponential for floating points
Up until now Gnat2Why incorrectly used the power function on reals of
Why3 for the exponential for floating points. This patch removes that
and leave the exponential for floating points as an uninterpreted
function. Additionally, an optimization is introduced that inline the
exponentiation when the exponent is known and is in {-3 .. +3}.

* _gnatprove_standard.mlw
  (Floating):
  The power function is now a non-interpreted function.
* gnat2why-expr.adb
  (Transform_Expr):
  Inline exponentiation when the exponent is in {-3 .. +3}.
* testsuite
  Some division checks from exponentiation disappears, which is ok
  since it happens when the exponentiation is inlined without a
  division.

Change-Id: I02eaf1f029835687f49d60af37b543dd142cd97e