P125-015 flow - check consistent refinement
In the case of a private child with Abstract_State annotated with Part_Of
we now make sure that there is consistence between this state and the
refinement in the relevant package body.

* gnat2why/flow/flow-analysis.ads
* gnat2why/flow/flow-analysis.adb
 Added procedure Check_Consistent_AS_For_Private_Child for the described
check.

* gnat2why/flow/flow.adb
(Flow_Analyse_CUnit): Call to Check_Consistent_AS_For_Private_Child

* gnat2why/flow/flow_generated_globals-phase2.ads
* gnat2why/flow/flow_generated_globals-phase2.adb
 Added function Find_In_Refinement that returns True iff a given
constituent is mentioned in the refinement of a given abstract state.

* testsuite/gnatprove/tests
Updated outputs