minor placeholder dir for non-native runtimes
Create a placeholder directory with a README file that explains to the
user what GNATprove expects to find there.

Previously users had to create this directory by themselves. Now we
prefer the approach taken by many Linux distributions, which is to
create placeholder directory with a README (or similar) file, e.g.
/etc/libpaper.d, /etc/rsyslog.d, /etc/sensors.d.