P706-032 Temporary fix for crashes related to new DIC handling
After FE changes on how DIC is handled, gnat2why fails to take into
account the special case of the parameter of the DIC procedure.
Special case this parameter temporarily, to fix crashes. We will need
to investigate why there is a change in this area.

* gnat2why/why/gnat2why-expr.adb
(Variables_In_Default_Init): Remove the parameter of the DIC procedure
  in the set being considered as a global variables.