P706-019 Update output after fix on renamings
Replacement of renaming by renamed object in GNATprove mode in the
frontend now uses correctly the location of the node being rewritten
instead of the source location of the renamed object inside the
renaming declaration. This fixes the crash reported in P706-019 and
leads to better locations for checks in one test.

* testsuite/gnatprove/tests/KA17-027__obj_renaming/test.out
Check is located on use of renaming instead of definition.

* testsuite/gnatprove/tests/P706-019__scopes_on_arm/test.opt
Test passes now.