P712-014 Restructure SPARK User's Guide to be able to create core document
A global MODE is defined which can take value 'full' (the default) or
'core' (which means a subset of the document is created). The core
document is ideally fit for translation in other languages. It is properly
cross-referenced and self-sufficient. To create the full document in HTML
for example, type:
     make html
and to create the core document in HTML, type:
     MODE=core make html

Also use now toctree as the main means to include ReST files, instead of
raw inclusion, which is a better way to include content.

* Makefile
Define a global MODE used to add a corresponding tag for building.

* conf.py
Change exclude_patterns now that inclusion is based on toctree instead of
raw include directive.

* gnatprove.rst
* spark_2014.rst
Separate out subsections in separate files, that are or not included in
the toctree depending on the mode.

* gnatprove_by_example/basic.rst
* gnatprove_by_example/loop.rst
* gnatprove_by_example/loop_patterns.rst
* gnatprove_by_example/manual_proof.rst
Changes from using toctree instead of raw include for inclusion.

* index.rst
Use mode to decide on which toctree to define.

* source/assertion_pragmas.rst
* source/concurrency.rst
* source/examples_in_the_toolset_distribution.rst
* source/gnatprove_by_example.rst
* source/how_to_investigate_unproved_checks.rst
* source/how_to_run_gnatprove.rst
* source/how_to_use_gnatprove_in_a_team.rst
* source/how_to_view_gnatprove_output.rst
* source/how_to_write_loop_invariants.rst
* source/how_to_write_object_oriented_contracts.rst
* source/how_to_write_package_contracts.rst
* source/how_to_write_subprogram_contracts.rst
* source/language_restrictions.rst
* source/object_oriented_programming.rst
* source/overflow_modes.rst
* source/package_contracts.rst
* source/spark_libraries.rst
* source/specification_features.rst
* source/subprogram_contracts.rst
* source/type_contracts.rst
New separated out subsections in their own files.