minor no special-case for anonymous task objects
For tasks the implicit parameter represents the task type itself, not the
anonymous task object.

This makes no difference for the graph, since both task type and its
anonymous objects are flattened to the same set of discriminants and Part_Of
variables.

* gnat2why/flow/flow-control_flow_graph.adb
(Create): always create vertices for the implicit parameter from the task
 type, not from the anonymous task object.