P627-026 Flatten_Variable for concurrent types
Cleanup in preparation for the real work: split code for record and
concurrent types and cleanup the latter.

For concurrent types now we flatten the discriminants, components (only in
protected types) and Part_Of variables (if visible); finally, we always
include the type itself, since from the outside it is all we see and from
the inside it represents the implicit "current instance" parameter (see
SPARK RM 6.4.1).

Previously we only included the implicit "current instance" if we did not
have anything better.

* gnat2why/flow/flow_utility.adb
(Flatten_Variable): split code for records and concurrent types