P524-016 flow - various small fixes
First, fixed a crash when checking for missing elaborate_body. In
particular, loop_parameters go into the set of local variables; but
really we must not be checking those in this case: only real variables
that are initialized should be checked.

This uncovered another issue with this check, we do not consider a nested
package to be affected by the pragma elaborate_body of its enclosing
package. This is now fixed.

And finally, a initialize (generated or explicit) is of the form A => B.
We used to a vertex using B and defining A to model this, but really if
A or B are records then we must flatten them. We did not do this, now we
do.

Enabled and updated tests.