P713-012 Printing the content of private types
We change the matching in Get_Counterexample_Labels to be able to match
private types as a normal kind because nodes corresponding to private
type definition are of kind private before being of their own type
(array, record or anything else). We proceed by taking the full view of
the private type which is a node that we can correctly consider
as a normal type.
Some tests were too complicated to be checked, but anything we print now
comes from a private type.

* gnat2why-util.adb
(Get_Counterexample_Labels): We use the function
Get_Full_Type_Without_Checking to recover the full_view of the type of
 E and then case on this instead of doing it on the type of E directly.

* L611-003__exist/test.out
  Update test
* L926-036__loopinv/test.out
  Update test
* M611-052__discr/test.out
  Update test. It seems that discriminants could be useful to check why
 the prover do not manage to prove it.
* N622-004__discr/test.out
  Update test
* NA11-009__very_long
  Update test
* NA11-009__very_long_subunit/test.out
  Update test
* O116-027__very_long/test.out
  Update test
* P503-001__bitsets/test.out
  Update test