Minor test cleanup
The test in question originates from Muen; updated code with the latest
version there (not much change, only types). The test is also a stripped
down version, and in particular the customer issue was in Mont_Exp_Window.
I have modified the test script to only attempt proof on that subprogram,
this reduces the run-time of the test a little bit.

Added a note to the test script that most of the time is actually spent
in gnatwhy3 and not in running cvc4.