P318-019 more hierarchy in flow traversal
Hierarchical traversal for global generation was respecting nesting in
packages; now it also respects nesting in functions, procedures, entries
(and entry families, which are not in SPARK), task types and protected
types.

There is a special-case for bodies of predicate functions (which perhaps
should extended to other delayed aspects of types), which may be put by the
front end in different lexical scopes than their specs.

Also, now we do not build graphs for subprograms whose body is in SPARK but
spec is not. This is still a workaround, for such entities should be not
even marked.

* gnat2why/spark/spark_definition.adb
(Scope_Trees): dedicated package for handling hierarchical scopes

* gnat2why/flow/flow.adb
(Build_Graphs_For_Entity): ignore subprograms whose spec is not in SPARK