O909-059 flow - detect missing part_of
We now check if there is any missing Part_Of indicator for non deferred
constants with variable input declared immediately within the private part
of a package or in the visible part of a package declared immediately
within the private part of another package. This needs to be done during
flow analysis because front-end cannot detect if a constant has variable
input or not.

* gnat2why/flow/flow-analysis-sanity.ads
* gnat2why/flow/flow-analysis-sanity.adb
(Check_Part_Of): added procedure for the check above described

* testsuite/gnatprove/tests/O909-059__part_of
Test filed