P613-015 Translate hidden record components
We use to not translate components hidden by a discriminant
constraint and to collapse components hidden by private
derivation along with components not in SPARK inside a field
named anc__.
We now translate every hidden field as a field in th why
record translation. This allows to refer to them without
needing a conversion. Fields that are not in SPARK are
modeled by a private field named after the type in which they
were declared.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Also store record fields for record subtypes.

* gnat2why/spark/spark_util-types.adb
(Get_Variant_Info): Renaming of Get_Component_Info.
(Component_Is_Visible_In_Type): Checks if a component is visible
in a type, that is, if it will be supplied by an aggregate.
(Get_Component_Set): Return the set of components of a record type.
Components that are not in SPARK are modeled by the type entity
in which they are declared.
(Original_Declaration): The first type in which a component appears.
(Search_Component_In_Type): Returns the component corresponding
to Comp in Rec. Should be used instead of Search_Component_By_Name,
as it also handles hidden fields.

* gnat2why/why/gnat2why-external_axioms.adb

(Compute_Substitution): Also renames record components if needed.

* gnat2why/why/why-atree-modules.adb
(Insert_Object_Symbols): Store symbols for component's
attributes in the record type's module. Use full name for them.

* gnat2why/why/why-gen-records.adb
(Declare_Rep_Record_Type): Remove __main and __anc fields for
partially private record derivation. Rather introduce a new
private component named rec__<Full_Name (Ty)> for each type
Ty in the derivation which has private fields of its own
(Has_Private_Fields). Also declare hidden record fields
which are in SPARK.
(New_Ada_Record_Aggregate): Generate a dummy association for
fields that are hidden in the type.

* gnat2why/why/why-inter.adb
(To_Why_Id): Use full name for record components as there may
be several with the same short name. The Rec parameter giving
the expected record type is now mandatory for record components
to tell the difference between fields for private types modeled
by types and actual types.

* share/spark/theories/ada__containers__
Use full name for record components.