P114-035 tool - improve posting SPARK violation
* gnat2why/spark/spark_definition.adb
(Mark_Function_Specification): we now post SPARK violations of the return
type of a function on the function entity itself.
(Mark_Type_Entity): in case of an array type we now post SPARK violations
of an index on the index itself rather than on the array type.

* testsuite/gnatprove/tests/*/test.out
Updated test ouput