P713-053 Fix handling of -gnatws for warnings issued after gnat2why
Some of the warnings issued by the frontend are issued after gnat2why
proper, inside the frontend. These should be controlled by -gnatws
rather than the value of --warnings. As a result, the changes to
Warning_Mode must be reversed after the call to the back-end, the counts
of warnings properly reset, and the presence of errors related to
warnings must be stored after each separate phase for a given value of
Warning_Mode.

* gnat2why/why/back_end.adb
(Call_Back_End): Save/restore Warning_Mode around call to back-end.
  Also store presence of errors at each phase.

* testsuite/gnatprove/tests/M603-021__volatile_input_proofs/test.out
* testsuite/gnatprove/tests/P201-069__simulink/test.out
* testsuite/gnatprove/tests/P713-053__warning/test.out
Update outputs given that -gnatws is given in the generated driver.