P721-003 flow - detect function in delay until statement
Flow analysis was not detecting function calls in a delay until statement.
Now this calls are included into the set of Functions_Called. This is
implemented with a new case for the node representing the expression of
the delay until statement. In particular if the node is of kind
N_Identifier and the its entity has kind E_Function, then the entity gets
included.

* gnat2why/flow/flow_utility.adb
(Collect_Functions_And_Read_Locked_POs.Proc): add case for N_Identifier
node whose entity kind is E_Function. Then include this entity into the set
of Functions_Called.

* testsuite/gnatprove/tests/P721-003__delay_function_call/test.out
Updated test output