P620-034 Fix unsound proof of floating-point range checks
When an actual parameter of a more constrained floating-point type than
the formal parameter is passed to a call, GNATprove issued a proved check
even when a run-time failure was possible. This was due to a misuse of
function Determine_Range_T which does not work for output actual
parameters.

* gnat2why/spark/spark_util.adb|s
(Is_Converted_Actual_Output_Parameter): New function to determine if
  an expression is an actual parameter possibly under conversion.

* gnat2why/why/why-gen-expr.adb
(Insert_Scalar_Conversion): Do not use interval arithmetic for an output
  actual parameter.

* testsuite/gnatprove/tests/P620-034__float_out_subtype/get_out_subtype.adb
* gnatprove/tests/P620-034__float_out_subtype/get_out_subtype_int.adb
* testsuite/gnatprove/tests/P620-034__float_out_subtype/test.out
Update test and output.