P718-014 Printing counterex in presence of return values
We encountered a problem with printing counterexemples value when an
application function is used. This is due to a transformation in why3
that introduce variables: introduce_premises. This transformation is used
in counterexmp generation. The problem was that introduction can occur
past the model_vc label and introducing counterexamples variables in the
context. This prevented us from retrieving variables in the goal.
This commit adds a label stop_intros which disallow the transformation to
introduce past it.

* gnat2why-subprograms.adb:
  (Generate_VCs_For_LSP)
  (Generate_VCs_For_Subprogram)
  Adding the right entity in the declaration of reference to function
 results as suggested by fumex
* gnat2why-util.adb:
  (Get_Counterexample_Labels) Special case if the entity is a function
because the return reference of a function should be model_projected
* vc_kinds.ads:
  Definition of Stop_Intro_Marker
* why-atree-modules.adb
* why-atree-modules.ads:
   Definition of the label stop_intro
* why-gen-expr.adb:
  (New_VC_Labels) adding the label stop_intro during generation of labels
for vc
* why3:
  see commit 730d3f1
Updated tests