P726-017 clean up calls to Retysp before marking
Retysp which is used to query for a type's representative
used to accept unmarked types. It was a no-op on them.
Now Retysp does not accept unmarked types except for
Itypes whose Retysp is the Retysp of their Etype.

* gnat2why/spark/spark_definition.adb
(Retysp_In_SPARK): New local function similar to In_SPARK
which marks the type entity before checking whether its
Retysp is in SPARK.
(Mark): Call Mark_Most_Underlying_Type before
Aggregate_Is_Fully_Initialized so that we do not query
for the Retysp of an unmarked type. Check the Retysp
of the node's Etype instead of the Etype directly.

* gnat2why/spark/spark_util-types.adb
(Retysp): Do not accept unmarked types except for
Itypes whose Retysp is the Retysp of their Etype.

* gnat2why/spark/spark_util.adb
(Root_Record_Component): If the Retysp of the component's
scope is root, search the component in the root type instead
of returning it unchanged as the type may not be its Retysp.