P727-001 do not mark record component that are not SPARK
We used to mark every component of a record type which is
not a private type. Somehow, with a tagged type declared
in a generic with a DIC, we end up with such a type which
has components that are not visible in SPARK. We should
not look at them while marking.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Only mark components of record types
which are visible in SPARK (not declared in a private part
with mode Off).