P629-001 Fix test and UG on correct use of unknown constants
GNATprove can be instructed to ignore the value of non-scalar constants
in proof, thus making it possible to perform parameterized proofs, by
using SPARK_Mode Off on the private part. Fix the test and SPARK UG to
indicate that this is not available for constants of scalar type (for which
the frontend may propagate the value when analyzing the AST).

* docs/ug/usage_scenarios.rst
Update UG.

* testsuite/gnatprove/tests/K228-006__deferred/client.adb
* testsuite/gnatprove/tests/K228-006__deferred/const.ads
* testsuite/gnatprove/tests/K228-006__deferred/test.out
Update test.