P218-017 support --proof-dir in --cache mode
Tests that include custom axioms (and thus run gnatwhy3 with --proof-dir
switch) did not gain any speed up when running the testsuite with the
--cache switch.

It was because the argument to --proof-dir is a full path to the test
sandbox and includes the PID of the test driver process, e.g.
/tmp/tmp-test-bitwalker-17123/src/proof_dir, so it is different for every
execution of the run-tests script.

Now we recognize the --proof-dir switch and do not hash the full path, but
instead walk the directory and cache the contents of the .mlw files. We skip
session and shapes files (and their .bak versions), because they are
uninteresting and fragile.

Also, for --socket switch now we ignore both the argument and the switch
itself. Previously only the argument was ignored but the switch was hashed,
so runs with and without why3server were cached as different. The same for
-j switch.

Finally, switches --debug and --force are now also ignored, since they
should not affect the result.

* scripts/cache_wrapper
(hash_proof_dir): new routine for hashing .mlw files in --proof-dir.
(hash_cmd_line): ignore more switches and their arguments.