P801-019 Fix crash on entry marked SPARK_Mode Off
When translating an entry whose body is marked SPARK_Mode Off in Why3,
we wrongly translated the entry guard, which is part of the entry body
and thus should not be translated when the body is SPARK_Mode Off.
This could result in a crash because special variable __self used for
the protected object was not available in the context.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_VCs_For_Subprogram): Do not translate the entry guard when the
  body is not in SPARK.

* testsuite/gnatprove/tests/P801-019__protected_off/pack.adb
* testsuite/gnatprove/tests/P801-019__protected_off/pack.ads
* testsuite/gnatprove/tests/P801-019__protected_off/test.out
* testsuite/gnatprove/tests/P801-019__protected_off/test.py
Add test.