minor revert to always generate Initializes
Revert 9b9471581030d6af41b0376ffbd67936d88b0e87, which introduced a diff on
O709-035__constant_with_var_input. That diff that was not detected while
checking the mailserver's results because the testsuite was not in really
good shape.

Apparently when globals hidden in a package are considered as implicit
abstract states we have to always generate Initializes aspect for them. It
is because they may be accessed by indirect calls paths.