P728-024 improve semantics of pragma Annotate
previusly, the pragma always applied to the entire range of the
construct to which it applied. Now, it applies to the range between the
statement and the pragma.

* gnat2why-annotate.adb:
(Insert_Annotate_Range): new argument whole, which applies the whole
range only if set to true, otherwise uses pragma as rightmost delimiter