P805-023 remove hierarchical scopes from marking
For new generated globals we need to traverse current compilation unit in a
hierarchical fashion. This was done by creating a hierarchical container
while marking (P318-019). However, this is fundamentally incompatible with
"lazy" marking which is the subject of this TN.

This patch reverts to semi-random traversal for flow analysis, but it makes
no difference until new GG is completed. Hierarchical traversal is still
kept in a local branch and nurtured with the rest of new GG.

* gnat2why/flow/flow.adb
(Build_Graphs_For_Compilation_Unit): revert to semi-random traversal of all
 marked entities.

(Build_Graphs_For_Entity): ignore uninteresting entities.

* gnat2why/spark/spark_definition.ads
(Iterate_Entities): remove entry point to hierarchical traversal.

* gnat2why/spark/spark_definition.adb
Remove all references to scope trees.