P820-005 don't check init status of remote states
When looking for procedures declared in a package that have its abstract
states as a pure output we should only care about abstract states belonging
to the very package that we analyse. This was assumed during "minor rewrite
in commit a25aa71b15211918973843c5dacf96873c6ce49a, but was not really the
case.

This patch fixes effects of that mistake but leaves the underlying problem
for later (as it only causes minor inefficiency) and flags it with ???.

* gnat2why/flow/flow-analysis.adb
(Collect_Procedure_Outputs): do not assume that an abstract state written by
 a procedure is an abstract state of the analyzed package; in effect, put it
 into containers with other abstract states, which is useless but does not
 harm.