minor no marking of pkg body if spec not in SPARK
If package specification is not in SPARK then there is no point in marking
package body. This save us both time and duplicated error messages.

Also, this cleans things up for the new GG where will not traverse bodies
of packages (and anything nested in there) if the spec is not in SPARK.

* gnat2why/spark/spark_definition.adb
(Mark_Package_Body): if package spec is not in SPARK then do not mark body.

* testsuite/gnatprove/tests/*/test.out
Test output updated: remove duplicated and not very useful messages.

xref K222-004, K302-029, N816-002, O429-045, O916-043.