P824-002 match front end and GG for generic subps
This patch complements the front end fix that provides unique entity names
for instances of generic subprograms both in front end cross-references and
generated globals.

The front end patch to Unique_Name introduced collisions between names of
instances of generic subprograms (which it handles as we would like) and
names of internal procedures created for default expressions (for which it
may wrongly return the name of instantiated procedure).

Fixing the name returned for these internal procedures is not worth the
effort: they are freezing artefacts that we should never care about, but we
did care not until now.

Note: we still ask for Unique_Name of these internal procedures (as can be
easily detected by adding an assertion in Unique_Name) in two naive
traversals: in front end when we collect SPARK cross-references and in
gnat2why when we call Register_Entity.

Change-Id: Ic1be7e79d283fb3dcaaefe99d8786329f331f205

* gnat2why/flow/flow_generated_globals-traversal.adb
(Process): ignore internal front end procedures.

* gnat2why/spark/spark_definition.adb
(Mark_Subprogram_Declaration): as above.

* testsuite/gnatprove/tests/euclidian_division/test.out
Test update after front end fix.